@@ -1,8 +1,8 @@
 <template lang="pug">
 	v-app
 		v-navigation-drawer.pb-0(persistent,light,:mini-variant.sync="mini",v-model="drawer")
-			v-list.pa-0
-				v-list-item
+			v-toolbar.transparent(flat)
+				v-list.pa-0
 					v-list-tile(avatar,tag="div")
 						v-list-tile-content
 							v-list-tile-title(v-if="SESSION.username") {{ SESSION.username }}
@@ -10,48 +10,47 @@
 						v-list-tile-action
 							v-btn(icon,@click.native.stop="mini = !mini")
 								v-icon chevron_left
-			v-list(dense)
+			v-list.pt-0(dense)
 				v-divider
-				v-list-item
-					v-list-tile(router=true,to="/login",v-if="! SESSION.authenticated")
-						v-list-tile-action
-							v-icon face
-						v-list-tile-content
-							v-list-tile-title {{ trans('app.menu.connect') }}
-				v-list-item
-					v-list-tile(router=true,to="/items",v-if="!! SESSION.authenticated")
-						v-list-tile-action
-							v-icon list
-						v-list-tile-content
-							v-list-tile-title {{ trans('app.menu.items') }}
-				v-list-item
-					v-list-tile(v-if="!! SESSION.authenticated",v-on:click.native="handleExport()")
-						v-list-tile-action
-							v-icon import_export
-						v-list-tile-content
-							v-list-tile-title {{ trans('app.menu.export') }}
+				v-list-tile(router=true,to="/login",v-if="! SESSION.authenticated")
+					v-list-tile-action
+						v-icon face
+					v-list-tile-content
+						v-list-tile-title {{ trans('app.menu.connect') }}
+				v-list-tile(router=true,to="/items",v-if="!! SESSION.authenticated")
+					v-list-tile-action
+						v-icon list
+					v-list-tile-content
+						v-list-tile-title {{ trans('app.menu.items') }}
+				v-list-tile(v-if="!! SESSION.authenticated",v-on:click.native="handleExport()")
+					v-list-tile-action
+						v-icon import_export
+					v-list-tile-content
+						v-list-tile-title {{ trans('app.menu.export') }}
 				v-divider
-				v-list-item
-					v-list-tile(v-if="!! SESSION.authenticated",v-on:click.native="handleLogout()")
-						v-list-tile-action
-							v-icon power_settings_new
-						v-list-tile-content
-							v-list-tile-title {{ trans('app.menu.logout') }}
-				v-list-item
-					v-list-tile(router=true,to="/about")
-						v-list-tile-action
-							v-icon chat_bubble
-						v-list-tile-content
-							v-list-tile-title {{ trans('app.menu.about') }}
+				v-list-tile(v-if="!! SESSION.authenticated",v-on:click.native="handleLogout()")
+					v-list-tile-action
+						v-icon power_settings_new
+					v-list-tile-content
+						v-list-tile-title {{ trans('app.menu.logout') }}
+				v-list-tile(router=true,to="/about")
+					v-list-tile-action
+						v-icon chat_bubble
+					v-list-tile-content
+						v-list-tile-title {{ trans('app.menu.about') }}
 
-		v-toolbar.indigo.darken-4(fixed,light)
-			v-toolbar-side-icon(light,@click.native.stop="drawer = !drawer")
-			v-toolbar-title.hidden-xs-only {{ trans('app.title') }}
-			v-text-field(prepend-icon="search",:label="trans('items:list.search')",v-on:input="search",hide-details,single-line,light)
+		v-toolbar.indigo.darken-4(fixed,dark)
+			v-toolbar-side-icon(@click.native.stop="drawer = !drawer")
+			v-toolbar-title {{ trans('app.title') }}
+
 		main
+			v-toolbar.indigo.darken-4(v-if="SESSION.authenticated")
+				v-text-field(solo,:label="trans('items:list.search')",v-on:input="search",prepend-icon="search")
+
 			v-container
 				router-view
-		v-footer.indigo.darken-4
+
+		v-footer.indigo.darken-4(dark)
 			span © 2017
 </template>
 
