@@ -1,88 +1,79 @@
 <template lang="pug">
 v-card
-	v-card-row.darken-1(v-bind:class="cardType.color")
-		v-card-title
-			span.white--text {{ trans(cardType.label) }}
-			v-spacer
-			v-btn.white--text(icon="icon",slot="activator",router=true,:to="'/items/' + line._id + '/edit'")
-				v-icon mode_edit
+	v-toolbar.white--text.darken-1(v-bind:class="cardType.color")
+		v-toolbar-title {{ trans(cardType.label) }}
+		v-spacer
+		v-btn.white--text(icon,router=true,:to="'/items/' + line._id + '/edit'")
+			v-icon mode_edit
 
 	v-list(two-line)
-		v-list-item
-			v-list-tile
-				v-list-tile-action
-					v-icon.indigo--text label
-				v-list-tile-content
-					v-list-tile-title {{ line.label }}
-					v-list-tile-sub-title {{ trans('items:item.form.label.field') }}
+		v-list-tile
+			v-list-tile-action
+				v-icon.indigo--text label
+			v-list-tile-content
+				v-list-tile-title {{ line.label }}
+				v-list-tile-sub-title {{ trans('items:item.form.label.field') }}
 
 	v-divider(inset)
 	v-list(three-line,v-if="line.type == 'text' && clearInformation.text")
-		v-list-item
-			v-list-tile
-				v-list-tile-action
-					v-icon.indigo--text text_fields
-				v-list-tile-content
-					v-list-tile-title {{ trans('items:item.form.text.field') }}
-					v-list-tile-sub-title {{ clearInformation.text }}
-				v-list-tile-action
-					v-icon(dark,v-on:click="copyToClipboard(clearInformation.text)") content_copy
+		v-list-tile
+			v-list-tile-action
+				v-icon.indigo--text text_fields
+			v-list-tile-content
+				v-list-tile-title {{ trans('items:item.form.text.field') }}
+				v-list-tile-sub-title {{ clearInformation.text }}
+			v-list-tile-action
+				v-icon(dark,v-on:click="copyToClipboard(clearInformation.text)") content_copy
 
 	template(v-if="line.type == 'card'")
 		v-list(two-line)
-			v-list-item(v-if="clearInformation.type")
-				v-list-tile
-					v-list-tile-action
-						v-icon.indigo--text credit_card
-					v-list-tile-content
-						v-list-tile-title {{ clearInformation.type }}
-						v-list-tile-sub-title {{ trans('items:item.form.type.field') }}
-			v-list-item(v-if="clearInformation.nameOnCard")
-				v-list-tile
-					v-list-tile-action
-						v-icon.indigo--text
-					v-list-tile-content
-						v-list-tile-title {{ clearInformation.nameOnCard }}
-						v-list-tile-sub-title {{ trans('items:item.form.nameOnCard.field') }}
-					v-list-tile-action
-						v-icon(dark,v-on:click="copyToClipboard(clearInformation.nameOnCard)") content_copy
-			v-list-item(v-if="clearInformation.cardNumber")
-				v-list-tile
-					v-list-tile-action
-						v-icon.indigo--text
-					v-list-tile-content
-						v-list-tile-title {{ clearInformation.cardNumber }}
-						v-list-tile-sub-title {{ trans('items:item.form.cardNumber.field') }}
-					v-list-tile-action
-						v-icon(dark,v-on:click="copyToClipboard(clearInformation.cardNumber)") content_copy
-			v-list-item(v-if="clearInformation.cvv")
-				v-list-tile
-					v-list-tile-action
-						v-icon.indigo--text
-					v-list-tile-content
-						v-list-tile-title {{ clearInformation.cvv }}
-						v-list-tile-sub-title {{ trans('items:item.form.cvv.field') }}
-					v-list-tile-action
-						v-icon(dark,v-on:click="copyToClipboard(clearInformation.cvv)") content_copy
-			v-list-item(v-if="clearInformation.expiry")
-				v-list-tile
-					v-list-tile-action
-						v-icon.indigo--text
-					v-list-tile-content
-						v-list-tile-title {{ clearInformation.expiry }}
-						v-list-tile-sub-title {{ trans('items:item.form.expiry.field') }}
-					v-list-tile-action
-						v-icon(dark,v-on:click="copyToClipboard(clearInformation.expiry)") content_copy
+			v-list-tile(v-if="clearInformation.type")
+				v-list-tile-action
+					v-icon.indigo--text credit_card
+				v-list-tile-content
+					v-list-tile-title {{ clearInformation.type }}
+					v-list-tile-sub-title {{ trans('items:item.form.type.field') }}
+			v-list-tile(v-if="clearInformation.nameOnCard")
+				v-list-tile-action
+					v-icon.indigo--text
+				v-list-tile-content
+					v-list-tile-title {{ clearInformation.nameOnCard }}
+					v-list-tile-sub-title {{ trans('items:item.form.nameOnCard.field') }}
+				v-list-tile-action
+					v-icon(dark,v-on:click="copyToClipboard(clearInformation.nameOnCard)") content_copy
+			v-list-tile(v-if="clearInformation.cardNumber")
+				v-list-tile-action
+					v-icon.indigo--text
+				v-list-tile-content
+					v-list-tile-title {{ clearInformation.cardNumber }}
+					v-list-tile-sub-title {{ trans('items:item.form.cardNumber.field') }}
+				v-list-tile-action
+					v-icon(dark,v-on:click="copyToClipboard(clearInformation.cardNumber)") content_copy
+			v-list-tile(v-if="clearInformation.cvv")
+				v-list-tile-action
+					v-icon.indigo--text
+				v-list-tile-content
+					v-list-tile-title {{ clearInformation.cvv }}
+					v-list-tile-sub-title {{ trans('items:item.form.cvv.field') }}
+				v-list-tile-action
+					v-icon(dark,v-on:click="copyToClipboard(clearInformation.cvv)") content_copy
+			v-list-tile(v-if="clearInformation.expiry")
+				v-list-tile-action
+					v-icon.indigo--text
+				v-list-tile-content
+					v-list-tile-title {{ clearInformation.expiry }}
+					v-list-tile-sub-title {{ trans('items:item.form.expiry.field') }}
+				v-list-tile-action
+					v-icon(dark,v-on:click="copyToClipboard(clearInformation.expiry)") content_copy
 
 		v-divider(inset)
 		v-list(two-line)
-			v-list-item(v-if="clearInformation.code")
-				v-list-tile
-					v-list-tile-action
-						v-icon.indigo--text vpn_key
-					v-list-tile-content
-						v-list-tile-title {{ clearInformation.code }}
-						v-list-tile-sub-title {{ trans('items:item.form.code.field') }}
+			v-list-tile(v-if="clearInformation.code")
+				v-list-tile-action
+					v-icon.indigo--text vpn_key
+				v-list-tile-content
+					v-list-tile-title {{ clearInformation.code }}
+					v-list-tile-sub-title {{ trans('items:item.form.code.field') }}
 
 	template(v-if="line.type == 'password'")
 		v-list(two-line)
