@@ -17,7 +17,7 @@ const networkInterface = createNetworkInterface({
 networkInterface.use([{
   applyMiddleware (req, next) {
     if (!req.options.headers) {
-      req.options.headers = {}  // Create the header object if needed.
+      req.options.headers = {} // Create the header object if needed.
     }
     req.options.headers = getAuthHeader()
     next()
